requirements-windows.txt
asciimatics==1.11.0
cffi
colorama==0.4.4
future==0.18.2
numpy==1.19.0
pash-cmd==0.0.7
Pillow==8.3.2
PyAudio==0.2.11
pycparser==2.20
pycryptodome==3.10.1
pydub==0.24.1
pyfiglet==0.8.post1
PyJWT==2.1.0
pynput==1.6.8
pypiwin32==223
python-xlib==0.29
pywin32==301
qrcode==6.1
six==1.15.0
wcwidth==0.2.5
websockets==9.1
Yapsy==1.12.2
